Add overlay behing editing content widget #274
