+++ client/mobilizations/widgets/__plugins__/content/components/editor-slate/index.scss
@@ -1,0 +1,3 @@
+.editor--content > div {
+  min-height: 370px;
+}
